refactor: reorganize project structure and package layout
