--- layout.schema.json
@@ -1,186 +1,0 @@
-{
-  "$defs": {
-    "TextField": {
-      "properties": {
-        "field": {
-          "title": "Field",
-          "type": "string"
-        },
-        "key": {
-          "title": "Key",
-          "type": "string"
-        }
-      },
-      "required": [
-        "field",
-        "key"
-      ],
-      "title": "TextField",
-      "type": "object"
-    }
-  },
-  "properties": {
-    "grid_shape": {
-      "description": "Grid size as [rows, columns], each >=1",
-      "maxItems": 2,
-      "minItems": 2,
-      "prefixItems": [
-        {
-          "minimum": 1,
-          "type": "integer"
-        },
-        {
-          "minimum": 1,
-          "type": "integer"
-        }
-      ],
-      "title": "Grid Shape",
-      "type": "array"
-    },
-    "font_list": {
-      "description": "Indexes of fonts to use",
-      "items": {
-        "type": "integer"
-      },
-      "minItems": 1,
-      "title": "Font List",
-      "type": "array"
-    },
-    "time_font": {
-      "description": "Index of font for time display",
-      "title": "Time Font",
-      "type": "integer"
-    },
-    "horizontal_spacing": {
-      "description": "Horizontal spacing in pixels",
-      "minimum": 0,
-      "title": "Horizontal Spacing",
-      "type": "integer"
-    },
-    "vertical_spacing": {
-      "description": "Vertical spacing in pixels",
-      "minimum": 0,
-      "title": "Vertical Spacing",
-      "type": "integer"
-    },
-    "content_margin_left": {
-      "description": "Left margin for content",
-      "minimum": 0,
-      "title": "Content Margin Left",
-      "type": "integer"
-    },
-    "content_margin_top": {
-      "description": "Top margin for content",
-      "minimum": 0,
-      "title": "Content Margin Top",
-      "type": "integer"
-    },
-    "title_margin_left": {
-      "description": "Left margin for title",
-      "minimum": 0,
-      "title": "Title Margin Left",
-      "type": "integer"
-    },
-    "title_margin_top": {
-      "description": "Top margin for title",
-      "minimum": 0,
-      "title": "Title Margin Top",
-      "type": "integer"
-    },
-    "shade_offset": {
-      "description": "Shadow offset values",
-      "maxItems": 2,
-      "minItems": 2,
-      "prefixItems": [
-        {
-          "type": "integer"
-        },
-        {
-          "type": "integer"
-        }
-      ],
-      "title": "Shade Offset",
-      "type": "array"
-    },
-    "text_color": {
-      "description": "RGB values for text color",
-      "items": {
-        "maximum": 255,
-        "minimum": 0,
-        "type": "integer"
-      },
-      "maxItems": 4,
-      "minItems": 3,
-      "title": "Text Color",
-      "type": "array"
-    },
-    "shade_color": {
-      "description": "RGB values for shadow color",
-      "items": {
-        "maximum": 255,
-        "minimum": 0,
-        "type": "integer"
-      },
-      "maxItems": 4,
-      "minItems": 3,
-      "title": "Shade Color",
-      "type": "array"
-    },
-    "text_list": {
-      "description": "Structured text entries for display",
-      "items": {
-        "items": {
-          "anyOf": [
-            {
-              "$ref": "#/$defs/TextField"
-            },
-            {
-              "type": "string"
-            }
-          ]
-        },
-        "type": "array"
-      },
-      "minItems": 1,
-      "title": "Text List",
-      "type": "array"
-    },
-    "pos_list": {
-      "description": "Position coordinates as [x, y]",
-      "items": {
-        "maxItems": 2,
-        "minItems": 2,
-        "prefixItems": [
-          {
-            "type": "integer"
-          },
-          {
-            "type": "integer"
-          }
-        ],
-        "type": "array"
-      },
-      "minItems": 1,
-      "title": "Pos List",
-      "type": "array"
-    }
-  },
-  "required": [
-    "grid_shape",
-    "font_list",
-    "time_font",
-    "horizontal_spacing",
-    "vertical_spacing",
-    "content_margin_left",
-    "content_margin_top",
-    "title_margin_left",
-    "title_margin_top",
-    "shade_offset",
-    "text_color",
-    "shade_color",
-    "text_list",
-    "pos_list"
-  ],
-  "title": "InfoLayout",
-  "type": "object"
-}
