+++ src/typings/__init__.py
@@ -1,0 +1,1 @@
+from .video_info import *
